y1apgm-django-multi-tenant-saas-api-with-row-level-security/repository_after/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
psycopg2-binary==2.9.9
